requirements.txt
attrs==20.3.0
click==7.1.2
flake8==3.9.0
Flask==1.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
toml==0.10.2
Werkzeug==1.0.1
flask-sqlalchemy
Flask-wtf
wtforms
psycopg2-binary
flask-bcrypt
Flask-Login
Flask-Script
Flask-Admin
Pycountry
intervals
SQLAlchemy-utils
faker
Pillow
babel
email-validator
flask-table
flask-mail
